メッセージをclient_implに分離
